pyproject.toml
[tool.poetry]
name = "o42-backend"
version = "0.1.0"
description = "Backend for o42 e-commerce platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
motor = "^2.5.1"
pydantic = "^1.8.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.1.0"
face_recognition = "^1.3.0"
sentence-transformers = "^2.0.0"
torch = "^1.9.0"
pillow = "^8.3.1"
aiosmtplib = "^1.1.6"
twilio = "^7.0.0"
geopy = "^2.2.0"
celery = "^5.1.2"
redis = "^3.5.3"
python-dotenv = "^0.19.0"
numpy = "1.26.4"
scipy = "1.11.4"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
httpx = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"